Se agrupan los data models en un mismo archivo
