+++ app/src/main/java/com/example/vinilosgrupo4/model/DataModels.kt
@@ -1,5 +1,15 @@
 package com.example.vinilosgrupo4.model
 
+data class MusiciansResponseDataModel (
+    val id: Long,
+    val name: String,
+    val image: String,
+    val description: String,
+    val birthDate: String,
+    val albums: List<Album>,
+    val performerPrizes: List<PerformerPrize>
+)
+
 data class BandReponseDataModel (
     val id: Long,
     val name: String,
