+++ app/src/main/java/com/example/vinilosgrupo4/view/adapter/itemsBandAdapter.kt
@@ -27,7 +27,6 @@ class itemsBandAdapter(private var listener: ClickBandListener): RecyclerView.Ad
         return itemBandViewHolder(binding)
     }
 
-
     override fun onBindViewHolder(holder: itemBandViewHolder, position: Int) {
         holder.setItem(itemList[position])
