+++ app/src/main/java/com/example/vinilosgrupo4/view/fragments/BandListFragment.kt
@@ -87,6 +87,10 @@ class BandListFragment : Fragment(), ClickBandListener {
     companion object {
         fun newInstance() = BandListFragment()
     }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+    }
 }
 
 interface ClickBandListener {
