+++ app/src/main/java/com/example/vinilosgrupo4/view/fragments/MusicianListFragment.kt
@@ -89,6 +89,10 @@ class MusicianListFragment : Fragment(), ClickMusicianListener {
     companion object {
         fun newInstance() = MusicianListFragment()
     }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+    }
 }
 
 interface ClickMusicianListener {
